fix: 유리수 속성 Long-> Double type 변경

+++ src/main/java/SOLog/SOLog/domain/entity/PriceDiffEntity.java
@@ -19,7 +19,7 @@ public class PriceDiffEntity {
     private Long id;
 
     private Date date;
-    private Long diff;
+    private Double diff;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "companyName")
